requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
chromadb==0.4.18
openai==1.3.5
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
textblob==0.17.1
python-dateutil==2.8.2
lxml==4.9.3
celery==5.3.4
redis==5.0.1
loguru==0.7.2